web/fileConfusion/src/app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pyyaml==5.1.2